feat: update rec no image
