+++ src/pymmcore_widgets/_stage_recorder.py
@@ -1,14 +1,13 @@
-from typing import Any, Optional
+from typing import Any, Optional, cast
 
 import numpy as np
 from fonticon_mdi6 import MDI6
 from pymmcore_plus import CMMCorePlus
-from qtpy.QtCore import QPoint, QSize, Qt, Signal
 from qtpy.QtWidgets import (
-    QCheckBox,
-    QDialog,
-    QHBoxLayout,
-    QPushButton,
+    QAction,
+    QMenu,
+    QToolBar,
+    QToolButton,
     QVBoxLayout,
     QWidget,
 )
@@ -16,70 +15,15 @@ from skimage.transform import resize
 from superqt.fonticon import icon
 from useq import MDAEvent
 from vispy import scene
-from vispy.color import Color
 from vispy.scene.visuals import Image
 from vispy.visuals.transforms import STTransform
 
-BTN_SIZE = (60, 40)
 SCALE_FACTOR = 3
-W = Color("white")
-G = Color("green")
-
-
-class _Settings(QDialog):
-
-    autoresetChanged = Signal(bool)
-
-    def __init__(self, parent: QWidget | None = None) -> None:
-        super().__init__(parent)
-
-        self.setWindowTitle("Settings")
-
-        combos_layout = QVBoxLayout()
-        combos_layout.setSpacing(10)
-        combos_layout.setContentsMargins(5, 5, 5, 5)
-        self.setLayout(combos_layout)
-        # flip horizontal checkbox
-        self._flip_horizontal_checkbox = QCheckBox("Flip Image Horizontally")
-        self._flip_horizontal_checkbox.setToolTip(
-            "Flip each acquired image horizontally."
-        )
-        self._flip_horizontal_checkbox.setChecked(False)
-        # flip vertical checkbox
-        self._flip_vertical_checkbox = QCheckBox("Flip Image Vertically")
-        self._flip_vertical_checkbox.setToolTip("Flip each acquired image vertically.")
-        self._flip_vertical_checkbox.setChecked(False)
-        # auto reset view checkbox
-        self._auto_reset_checkbox = QCheckBox("Auto Reset View")
-        self._auto_reset_checkbox.setToolTip("Automatically reset the view.")
-        self._auto_reset_checkbox.setChecked(True)
-        self._auto_reset_checkbox.toggled.connect(self._on_auto_reset_toggle)
-        # autosnap checkbox
-        self._autosnap_checkbox = QCheckBox("Auto Snap on double click")
-        self._autosnap_checkbox.setToolTip(
-            "Automatically snap an image when double clicking on the view."
-        )
-        self._autosnap_checkbox.setChecked(False)
-
-        combos_layout.addWidget(self._auto_reset_checkbox)
-        combos_layout.addWidget(self._autosnap_checkbox)
-        combos_layout.addWidget(self._flip_horizontal_checkbox)
-        combos_layout.addWidget(self._flip_vertical_checkbox)
-
-    def _on_auto_reset_toggle(self, state: bool) -> None:
-        self.autoresetChanged.emit(state)
-
-    def auto_reset(self) -> bool:
-        return bool(self._auto_reset_checkbox.isChecked())
-
-    def autosnap(self) -> bool:
-        return bool(self._autosnap_checkbox.isChecked())
-
-    def flip_horizontal(self) -> bool:
-        return bool(self._flip_horizontal_checkbox.isChecked())
-
-    def flip_vertical(self) -> bool:
-        return bool(self._flip_vertical_checkbox.isChecked())
+GRAY = "#666"
+RESET = "Auto Reset View"
+SNAP = "Auto Snap on double click"
+FLIP_H = "Flip Image Horizontally"
+FLIP_V = "Flip Image Vertically"
 
 
 class StageRecorder(QWidget):
@@ -95,75 +39,78 @@ class StageRecorder(QWidget):
 
         self._mmc = mmcore or CMMCorePlus.instance()
 
-        self._settings_menu = _Settings(self)
-        # Set window flags to make it unmovable and stay on top
-        self._settings_menu.setWindowFlags(
-            Qt.WindowType.Dialog | Qt.WindowType.WindowStaysOnTopHint
-        )
-
         self._visited_positions: list[tuple[float, float]] = []
         self._fov_max: tuple[float, float] = (0, 0)
+        self._auto_reset: bool = True
+        self._snap: bool = False
+        self._flip_h: bool = False
+        self._flip_v: bool = False
 
         # canvas and view
         self.canvas = scene.SceneCanvas(keys="interactive", show=True)
         self.view = self.canvas.central_widget.add_view()
         self.view.camera = scene.PanZoomCamera(aspect=1)
 
-        # buttons
-        btns = QWidget()
-        btns_layout = QHBoxLayout()
-        btns_layout.setSpacing(10)
-        btns_layout.setContentsMargins(5, 5, 5, 5)
-        btns.setLayout(btns_layout)
-        # clear button
-        self._clear_btn = QPushButton()
-        self._clear_btn.setToolTip("Clear the view.")
-        self._clear_btn.setFocusPolicy(Qt.FocusPolicy.NoFocus)
-        self._clear_btn.setIcon(icon(MDI6.close_box_outline))
-        self._clear_btn.setIconSize(QSize(25, 25))
-        self._clear_btn.setFixedSize(*BTN_SIZE)
-        self._clear_btn.clicked.connect(self.clear)
-        # reset view button
-        self._reset_view_btn = QPushButton()
-        self._reset_view_btn.setToolTip("Reset the view.")
-        self._reset_view_btn.setFocusPolicy(Qt.FocusPolicy.NoFocus)
-        self._reset_view_btn.setIcon(icon(MDI6.home_outline))
-        self._reset_view_btn.setIconSize(QSize(25, 25))
-        self._reset_view_btn.setFixedSize(*BTN_SIZE)
-        self._reset_view_btn.clicked.connect(self.reset_view)
-        # settings buttons
-        self._settings_btn = QPushButton()
-        self._settings_btn.setToolTip("Open settings menu.")
-        self._settings_btn.setFocusPolicy(Qt.FocusPolicy.NoFocus)
-        self._settings_btn.setIcon(icon(MDI6.cog_outline))
-        self._settings_btn.setIconSize(QSize(25, 25))
-        self._settings_btn.setFixedSize(*BTN_SIZE)
-        self._settings_btn.clicked.connect(self._show_settings_menu)
-        # add buttons to layout
-        btns_layout.addStretch(1)
-        btns_layout.addWidget(self._clear_btn)
-        btns_layout.addWidget(self._reset_view_btn)
-        btns_layout.addWidget(self._settings_btn)
-
         # add to main layout
         main_layout = QVBoxLayout()
         main_layout.setSpacing(0)
         main_layout.setContentsMargins(5, 5, 5, 5)
         self.setLayout(main_layout)
-        # main_layout.addWidget(combos)
-        main_layout.addWidget(self.canvas.native)
-        main_layout.addWidget(btns)
 
-        # this is to make the widget square
-        # self.setMinimumHeight(self.minimumSizeHint().width())
+        # toolbar
+        toolbar = QToolBar()
+        toolbar.setMovable(False)
+
+        # reset view action
+        self.act_reset_view = QAction(
+            icon(MDI6.home_outline, color=GRAY), "Reset View", self
+        )
+        self.act_reset_view.triggered.connect(self.reset_view)
+        toolbar.addAction(self.act_reset_view)
+
+        # clear action
+        self.act_clear = QAction(
+            icon(MDI6.close_box_outline, color=GRAY), "Clear View", self
+        )
+        self.act_clear.triggered.connect(self.clear_view)
+        toolbar.addAction(self.act_clear)
+
+        # settings button and context menu
+        # create settings button
+        self._settings_btn = QToolButton()
+        self._settings_btn.setPopupMode(QToolButton.ToolButtonPopupMode.InstantPopup)
+        self._settings_btn.setToolTip("Settings Menu")
+        self._settings_btn.setIcon(icon(MDI6.cog_outline, color=GRAY))
+        toolbar.addWidget(self._settings_btn)
+        # create context menu
+        menu = QMenu(self)
+        # connect the menu to the button click
+        self._settings_btn.setMenu(menu)
+        # create actions for checkboxes
+        auto_reset_act = QAction(RESET, self, checkable=True, checked=True)
+        auto_snap_act = QAction(SNAP, self, checkable=True)
+        flip_h_act = QAction(FLIP_H, self, checkable=True)
+        flip_v_act = QAction(FLIP_V, self, checkable=True)
+        # add actions to the menu
+        menu.addAction(auto_reset_act)
+        menu.addAction(auto_snap_act)
+        menu.addAction(flip_h_act)
+        menu.addAction(flip_v_act)
+        # add actions to the checkboxes if needed
+        auto_reset_act.triggered.connect(self._on_setting_checked)
+        auto_snap_act.triggered.connect(self._on_setting_checked)
+        flip_h_act.triggered.connect(self._on_setting_checked)
+        flip_v_act.triggered.connect(self._on_setting_checked)
+
+        # add to main layout
+        main_layout.addWidget(toolbar)
+        main_layout.addWidget(self.canvas.native)
 
         # connect signals
         self._mmc.events.imageSnapped.connect(self._on_image_snapped)
         self._mmc.mda.events.frameReady.connect(self._on_frame_ready)
 
         self.canvas.events.mouse_double_click.connect(self._on_mouse_double_click)
-
-        self._settings_menu.autoresetChanged.connect(self._on_reset_view_toggle)
 
     def value(self) -> list[tuple[float, float]]:
         """Return the visited positions."""
@@ -172,11 +119,24 @@ class StageRecorder(QWidget):
             (x * SCALE_FACTOR, y * SCALE_FACTOR) for x, y in self._visited_positions
         ]
 
+    def _on_setting_checked(self, checked: bool) -> None:
+        """Handle the settings checkboxes."""
+        sender = cast(QAction, self.sender()).text()
+        if sender == RESET:
+            self._auto_reset = checked
+        elif sender == SNAP:
+            self._snap = checked
+        elif sender == FLIP_H:
+            self._flip_h = checked
+        elif sender == FLIP_V:
+            self._flip_v = checked
+
     def _on_mouse_double_click(self, event: Any) -> None:
         """Move the stage to the mouse position.
 
         If the autosnap checkbox is checked, also snap an image.
         """
+        # if the mda is running, return
         if self._mmc.mda.is_running():
             return
 
@@ -187,29 +147,8 @@ class StageRecorder(QWidget):
             x, y, _, _ = self.view.camera.transform.imap(event.pos)
             self._mmc.setXYPosition(x * SCALE_FACTOR, y * SCALE_FACTOR)
 
-        if self._settings_menu.autosnap() and not self._mmc.isSequenceRunning():
+        if self._snap and not self._mmc.isSequenceRunning():
             self._mmc.snapImage()
-
-    def _show_settings_menu(self) -> None:
-        """Show the settings menu."""
-        # get the position of the settings button
-        pos = self._settings_btn.mapToGlobal(QPoint(0, 0))
-        # get the width and height of the settings button
-        button_width = self._settings_btn.width()
-        button_height = self._settings_btn.height()
-        # show the settings menu and move it to the top edge of the settings button
-        self._settings_menu.move(
-            pos.x() - (button_width * 2), pos.y() - (button_height * 3)
-        )
-        # hide the menu if it is already visible, otherwise show it
-        if self._settings_menu.isVisible():
-            self._settings_menu.hide()
-        else:
-            self._settings_menu.show()
-
-    def _on_reset_view_toggle(self, state: bool) -> None:
-        if state:
-            self.reset_view()
 
     def _set_max_fov(self) -> None:
         """Set the max fov based on the image size.
@@ -263,7 +202,7 @@ class StageRecorder(QWidget):
             if isinstance(child, Image):
                 child.parent = None
 
-    def clear(self) -> None:
+    def clear_view(self) -> None:
         """Clear the scene and the visited positions."""
         # clear visited position list
         self._visited_positions.clear()
@@ -310,10 +249,10 @@ class StageRecorder(QWidget):
         # scale
         scaled = resize(image, (height, width))
         # flip horizontally
-        if self._settings_menu.flip_horizontal():
+        if self._flip_h:
             scaled = np.fliplr(scaled)
         # flip vertically
-        if self._settings_menu.flip_vertical():
+        if self._flip_v:
             scaled = np.flipud(scaled)
 
         scaled_8bit = (scaled / scaled.max()) * 255
@@ -329,5 +268,5 @@ class StageRecorder(QWidget):
         y -= height / 2
         frame.transform = STTransform(translate=(x, y))
 
-        if self._settings_menu.auto_reset():
+        if self._auto_reset:
             self.reset_view()
